feat: epaxos: add grpc service definition and test of basice service
I'd like to rename the replication algorithm to Q-paxos, Q for quick or
cute or quantum :).

The impl of replication server does nothing but responding an empty body.

The only test added in this commit is to run a server in the same
process and send a request to it.