refactor: simplify Storage structure.
Storage now contains a Arc<dyn RawKV> as underlying shared storage
engine. NsStorage is removed.

Storage impl namespace. A replica has a dedicate Storage instance for
it.
RawKV does not provides namespace. RocksDB or MemEngine is impl as a
RawKV.

Rename trait ToKey to AsStorageKey, which defines API for an struct to
be used as key in a KV storage.