Add proto Record to store the value part of a record
Add proto Record to store the value part of a record. Besides the
serialized value, there will be other info such as the last instance
that update this value etc.

Remove From<> impl from WriteEntry. Because batch-write has no namespace
support thus a WriteEntry is Storage-related. Now building a WriteEntry
is reponsibility of a Storage instance.

Fix a bug that replica uses a storage with two layer of namespace
wrapping.